[Feature/#34] : 나머지 api 연결

+++ app/src/main/java/com/hackathon/alddeul_babsang/data/datasource/ProfileDataSource.kt
@@ -6,7 +6,7 @@ import com.hackathon.alddeul_babsang.data.dto.response.ResponseLikesDto
 
 interface ProfileDataSource {
     suspend fun getLikes(
-        userId: Long
+        userId: Int
     ): BaseResponse<ResponseLikesDto>
 
     suspend fun postLike(
